init pages and linked in navs:
Each weather, about and examples components are created respectively.
After, each components are embedded inside app.jsx. In the case of
Weather component, as it is the main(index) component it is embedded as
indexroute. In the Nav component, Link is called from react-router and
links to each page are created.