update Link in Nav:
Each Link now has activeClassName which is giving a html class when the
link is active. Also with activeStyle, you can give css syle to the
active class. Furthermore, when router is nested as <IndexRouter>, then
it is more ideal to use <IndexLink> rather than <Link>. Unless, it will
always match the url, so it would be always activated.