update weather components:
Once the form is submitted, it passes the value to the weather
component. Then triggers handleSearch which is intend to fetch the
weather information form the api but for now just set the temperature as
23. So the state is designated(with initial state and new setstate), the
value is passed in weather message component as location and temp
respectively. Once the Message component takes new state value, which is
props in Form component itself, it rerenders the message to new value.