update navbar:
By using foundation's styling classes, html elemets and classNames are
included in Navs. The links to each page is added on the left side, and
weather searching bar is included on the right pane of the nav.
Currently this search bar only triggers simple alarm. The genuine search
function is not wired up yet.