update examples and main components:
In Main.jsx, new div with foundation styling is added. All the
props.children is going under the styled div which centers the content
on the page.

In the Examples.jsx, h1 is styled, Link of react-router is added to link
to specific url. In here, the links are quering the temperature of
location. The function is not created yet.