update Weather.jsx: to handle modal
ErrorModal.jsx component is created and imported in Weather.jsx. In
Weather component, a new state called errorMessage, which handles if the
error message exists or not.

Inside Weather componenet, the handleSearch callback function is
changing this state. It the openWeatherMap can't bring proper data, it
opens up error and update the errorMessage state to event's message. But
every time the method is activated, this state is set to undefined to
clear it.

Inside the render, the conditional checks if the errorMessge state is
false(doesn't exist) or true(error happened). If it is true, it renders
the ErrorModal component inside the <div>.

Since the ErrorModal.jsx component has attribute componentDidMount which
is activated when the component is literally 'mounted' in another
component, when it is rendered the modal automatically opens up because
I set it to be open with modal.open().