update css: switched to .scss
For using scss preprocessor, the former css file is changed into .scss
file. I installed some sass and scss related npm packages. Now in the
scss file, the values can be saved as variable. This is so fucking
awesome!!! I should've used this way before... Gosh... Anyways to make
sure this works, in the app.jsx add 'sass!' in the require field to
allow webpack loader to process the scss.