update Weather.jsx: take query from url and search
In Weather.jsx, componentDidMount method is created which is activated
when the current component is "rendered" or "mounted" to current DOM.

Once the query is added in the url and this component is mounted, it
gets location of the query by using react-router. It gets it with
`this.props.location.query`. Then save the value in the variable
location, and start running search method 'handleSearch' by passing in
the value location.