update query handled search:
There was a problem after the url query search is completed, the query
still remains in the url. So when the page is refreshed it still process
the query searching. So by adding this syntax, right after the search
process, it emties the current url.

The window.location is about the current location of the document.