update navbar searchbox querying:
In navbar, the value input in searchbox is taken and encoded in URI
component then the value is passed as new url(which triggers Weather.jsx
component on the DOM).

But this only works when this search is made from other component page
as componentDidMount function works.
When user tries to search through navbar at the Weather.jsx component,
it doesn't work because the DOM is already mounted, componentDidMount
obviously don't work.

So I used componentWillReceiveProps function. This allows react to
rerender the mounted component when new props are received.